calc-back/requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.9
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
python-dateutil==2.8.1
pytz==2019.2
six==1.14.0
Werkzeug==1.0.0
flask-restful
Flask-Migrate
Flask-HTTPAuth
Flask-SQLAlchemy
marshmallow_enum
numpy
flask-bcrypt
flask-mail
numpy-financial
flask-sqlalchemy
python-dotenv==0.15.0
metomi-isodatetime
pymysql
